+++ ESLint.novaextension/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## future
+
+### Added
+
+- Support untitled (new, unsaved) documents
+
+### Changed
+
+- When fixing all issues, apply fixes atomically in editor before saving (make's them must faster)
+
+### Fixed
+
+- Fixed linting failing when extension is first activated
+- Automatically re-lint when preferences change
+
 ## v1.3.0
 
 ### Added